# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Phonora pronunciation coaching API
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardSummaryTopPracticedSoundsItem } from "./dashboardSummaryTopPracticedSoundsItem";

export interface DashboardSummary {
  streakDays: number;
  totalAttempts: number;
  todayAttempts: number;
  weekAttempts: number;
  /** @nullable */
  avgScoreWeek?: number | null;
  /** @nullable */
  avgScoreAllTime?: number | null;
  topPracticedSounds: DashboardSummaryTopPracticedSoundsItem[];
  /** @nullable */
  lastPracticeAt?: Date | null;
}