# lib/api-zod/src/generated/types/lookupPronunciationBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Phonora pronunciation coaching API
 * OpenAPI spec version: 0.1.0
 */

export interface LookupPronunciationBody {
  text: string;
}