# lib/api-zod/src/generated/types/practiceAttempt.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Phonora pronunciation coaching API
 * OpenAPI spec version: 0.1.0
 */
import type { PracticeAttemptLikelySubstitutionsItem } from "./practiceAttemptLikelySubstitutionsItem";
import type { PracticeAttemptPhonemeFeedbackItem } from "./practiceAttemptPhonemeFeedbackItem";

export interface PracticeAttempt {
  id: number;
  userId: string;
  practiceText: string;
  practiceType: string;
  /** @nullable */
  overallScore?: number | null;
  /** @nullable */
  accuracyScore?: number | null;
  /** @nullable */
  fluencyScore?: number | null;
  /** @nullable */
  completenessScore?: number | null;
  /** @nullable */
  transcript?: string | null;
  likelySubstitutions: PracticeAttemptLikelySubstitutionsItem[];
  phonemeFeedback: PracticeAttemptPhonemeFeedbackItem[];
  articulatoryTips: string[];
  /** @nullable */
  ipaBreakdown?: string | null;
  /** @nullable */
  durationMs?: number | null;
  createdAt: Date;
}