# lib/api-zod/src/generated/types/savedText.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Phonora pronunciation coaching API
 * OpenAPI spec version: 0.1.0
 */

export interface SavedText {
  id: number;
  userId: string;
  label: string;
  text: string;
  practiceType: string;
  createdAt: Date;
}