# lib/api-zod/src/generated/types/wordPronunciation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Phonora pronunciation coaching API
 * OpenAPI spec version: 0.1.0
 */

export interface WordPronunciation {
  word: string;
  pronunciations: string[];
  found: boolean;
}